Fix FORCE_NON_DEBUG_XEN symlink swizzling
The new logic tested for the existence of "/boot/${NON_DEBUG_XEN}". But if that variable was the empty string (as in pre-Xen 4.3 builds that did not include a debug Xen) this test would erroneously succeed, leading to /boot/xen.gz being deleted and the host failing to boot.

+++ exec/xenrt/lib/xenserver/host.py
@@ -1451,7 +1451,7 @@ then
     NON_DEBUG_XEN=$(basename $(readlink boot/xen-debug.gz) -d.gz).gz
 fi
 
-if [ -e "boot/${NON_DEBUG_XEN}" ]
+if [ -n "${NON_DEBUG_XEN}" -a -e "boot/${NON_DEBUG_XEN}" ]
 then
     rm -f boot/xen.gz
     ln -s ${NON_DEBUG_XEN} boot/xen.gz
